Fix module version, switch to ncc
Having node_modules in the repo is problematic. For one thing,
dependabot updates dependencies in package.json, but the old versions of
modules are still there in node_modules.

To fix this, let's switch away from having node_modules in the repo to
using ncc, as they recommend in [1] and [2].

Fix package version in package.json while we're at it.

This commit is mostly auto-generated, the only manual edited files are
package.json and .gitignore. Here is the workflow (using npm 8.9.12):

        $EDITOR package.json
        $EDITOR .gitignore
        git add .gitignore
        git rm -rf node_modules
        rm package-lock.json
        npm i
        git add dist
        git commit

NOTE that this also upgrades @actions/core to the latest 1.x release
(which is 1.10.x at the moment), which probably fixes the issue of GHA
warning about the use of deprecated set-output command. This happens
because ^1.0.9 means "latest 1.x" in npm.

[1] https://github.com/actions/javascript-action/tree/main
[2] https://docs.github.com/en/actions/creating-actions/creating-a-javascript-action

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>